+++ src/Commands/Tag.php
@@ -18,7 +18,7 @@ class Tag extends Command
      *
      * @var string
      */
-    protected $description = "[Git][Tag] create tag and push to repository";
+    protected $description = "Create tag and push to repository";
 
     /**
      * Create a new command instance.
@@ -37,28 +37,37 @@ class Tag extends Command
      */
     public function handle()
     {
-        $tagname = $this->ask('Tag version [and press enter]', '');
-
+        $tagname = $this->ask('Tag version [and press enter]', 'versión 0.0.1');
         $this->greetings();
-
-        if($tagname == ''){
-            $this->alert('~# Need a tagname to push');
-        }else{
-            exec('git tag -a ' . $tagname . ' -m "Versión ' . $tagname . '"');
-            exec('git push origin '.$tagname);
+        $option = $this->argument('option', '');
+        switch ($option) {
+            case 'create':
+                exec('git tag -a ' . $tagname . ' -m "Versión ' . $tagname . '"');
+                exec('git push origin ' . $tagname);
+                break;
+            case 'delete':
+                exec('git tag -d ' . $tagname);
+                exec('git push origin :refs/tags/' . $tagname);
+                break;
+            default:
+                $this->alert('=> option not exist');
+                break;
         }
-
         $this->farewell();
     }
 
     public function greetings()
     {
-        $this->info('~# Init Command');
+        $this->info('#############');
+        $this->info('Init Command');
+        $this->info('-------------');
     }
 
     public function farewell()
     {
-        $this->info('~# End Command');
+        $this->info('-------------');
+        $this->info('End Command');
+        $this->info('#############');
     }
 
 }