+++ src/basherServiceProvider.php
@@ -19,6 +19,7 @@ class basherServiceProvider extends ServiceProvider
         'fitluismacedo\basher\Commands\Revert',
         'fitluismacedo\basher\Commands\Tag',
         'fitluismacedo\basher\Commands\Ignore',
+        'fitluismacedo\basher\Commands\LogIt',
     ];
     /**
      * Perform post-registration booting of services.